requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
dgl-cu101==0.4.3.post1
idna==2.9
joblib==0.14.1
jsonpatch==1.25
jsonpointer==2.0
kiwisolver==1.0.1
matplotlib==3.1.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
natsort==7.0.1
networkx==2.4
numpy==1.18.1
olefile==0.46
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
pyzmq==19.0.0
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
sip==4.19.13
six==1.14.0
torch==1.4.0
torchfile==0.1.0
torchvision==0.5.0
tornado==6.0.4
urllib3==1.25.8
visdom==0.1.8.9
websocket-client==0.57.0
wincertstore==0.2
